Refactored conversion package
