+++ xlwings/conversion/framework.py
@@ -1,0 +1,134 @@
+# -*- coding: utf-8 -*-
+
+
+
+class ConversionContext(object):
+    __slots__ = ['range', 'value']
+
+    def __init__(self, range=None, value=None):
+        self.range = range
+        self.value = value
+
+
+class Options(dict):
+
+    def __init__(self, original):
+        super(Options, self).__init__(original)
+
+    def override(self, **overrides):
+        self.update(overrides)
+        return self
+
+    def erase(self, keys):
+        for key in keys:
+            self.pop(key, None)
+        return self
+
+    def defaults(self, **defaults):
+        for k, v in defaults.items():
+            self.setdefault(k, v)
+        return self
+
+
+class Pipeline(list):
+
+    def prepend_stage(self, stage, only_if=True):
+        if only_if:
+            self.insert(0, stage)
+        return self
+
+    def append_stage(self, stage, only_if=True):
+        if only_if:
+            self.append(stage)
+        return self
+
+    def insert_stage(self, stage, index=None, after=None, before=None, replace=None, only_if=True):
+        if only_if:
+            if sum(x is not None for x in (index, after, before, replace)) != 1:
+                raise ValueError("Must specify exactly one of arguments: index, after, before, replace")
+            if index is not None:
+                indices = (index,)
+            elif after is not None:
+                indices = tuple(i+1 for i, x in enumerate(self) if isinstance(x, after))
+            elif before is not None:
+                indices = tuple(i for i, x in enumerate(self) if isinstance(x, before))
+            elif replace is not None:
+                for i, x in enumerate(self):
+                    if isinstance(x, replace):
+                        self[i] = stage
+                return self
+            for i in reversed(indices):
+                self.insert(i, stage)
+        return self
+
+    def __call__(self, *args, **kwargs):
+        for stage in self:
+            stage(*args, **kwargs)
+
+
+converters = {}
+
+
+class Accessor(object):
+
+    @classmethod
+    def install_for(cls, *types):
+        for type in types:
+            converters[type] = cls
+
+    @classmethod
+    def router(cls, value, rng, options):
+        return converters.get(type(value), converters[None])
+
+
+class ConverterAccessor(Accessor):
+
+    class ToValueStage(object):
+
+        def __init__(self, write_value, options):
+            self.write_value = write_value
+            self.options = options
+
+        def __call__(self, c):
+            c.value = self.write_value(c.value, self.options)
+
+    class FromValueStage(object):
+
+        def __init__(self, read_value, options):
+            self.read_value = read_value
+            self.options = options
+
+        def __call__(self, c):
+            c.value = self.read_value(c.value, self.options)
+
+    base_type = None
+
+    @classmethod
+    def base_reader(cls, options, base_type=None):
+        return converters[base_type or cls.base_type].reader(options)
+
+    @classmethod
+    def base_writer(cls, options, base_type=None):
+        return converters[base_type or cls.base_type].writer(options)
+
+    @classmethod
+    def reader(cls, options):
+        return (
+            cls.base_reader(options)
+            .append_stage(ConverterAccessor.FromValueStage(cls.read_value, options))
+        )
+
+    @classmethod
+    def writer(cls, options):
+        return (
+            cls.base_writer(options)
+            .prepend_stage(ConverterAccessor.ToValueStage(cls.write_value, options))
+        )
+
+    @classmethod
+    def router(cls, value, rng, options):
+        if isinstance(value, cls.writes_types):
+            return cls
+        else:
+            return super(ConverterAccessor, cls).router(value, rng, options)
+
